% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_cleanup.R
\name{CT_cleanup}
\alias{CT_cleanup}
\title{Import, clean up, and export multiple raw conductivity-temperature, CT, data files from the HOBO CT Logger}
\usage{
CT_cleanup(
  data.path,
  output.path,
  path.pattern,
  tf.write = FALSE,
  tf.recursive = FALSE,
  hobo.cal = FALSE,
  cond.range = "high"
)
}
\arguments{
\item{data.path}{Path to the input files}

\item{output.path}{Path for the output files}

\item{path.pattern}{Character string or object referencing a character string that identifies CT logger data within the specified file path}

\item{tf.write}{Logical parameter indicating whether to save output files in an output folder. Default = FALSE.}

\item{tf.recursive}{Logical parameter indicating whether to search within folders at the data.path. Default = FALSE}

\item{hobo.cal}{Logical parameter indicating whether the .hobo file has been pre-calibrated using HOBOware software and exported as a processed .csv. Default = FALSE}

\item{cond.range}{Character string or object indicating whether conductivity was recorded in High ("high"), Low ("low"), or both ("highlow") ranges. Default = "high"}
}
\value{
For every imported CT data file, one tidied file with temperature-compensated conductance is exported and returned
}
\description{
This function loads all raw .csv files within the data.path folder.
It removes the first row of each file containing logger information, then retains column headers.
Unnecessary columns are removed, and the remaining columns are renamed as needed for simplicity and consistency.
Any rows containing NA's are dropped.
}
