% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_one_cal.R
\name{CT_one_cal}
\alias{CT_one_cal}
\title{Calculate temperature-compensated Specific Conductance and one-point calibration for CT logger data}
\usage{
CT_one_cal(data, date, temp, EC, cal_reference, startCal, endCal)
}
\arguments{
\item{data}{The dataframe containing raw Electrical Conductivity values}

\item{date}{Date column used to filter for calibration data only}

\item{temp}{Temperature column used to calculate specific conductance from electrical conductivity}

\item{EC}{Electrical conductivity column used to calculate specific conductance}

\item{cal_reference}{The conductivity calibration solution specific conductance value, uS/cm at 25degC}

\item{startCal}{Date and time at the start of the calibration}

\item{endCal}{Date and time at the end of the calibration}
}
\value{
The original dataframe with the newly calculated Specific Conductance and calibrated SC values of CT logger data
}
\description{
This function reads in a dataframe containing electrical conductivity logs from a conductivity-temperature logger
and calculates specific conductance from in situ electrical conductivity and temperature readings.
}
\details{
Nonlinear temperature compensation is calculated using the EC to SC equation from
https://www.aqion.de/site/112
}
