% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_roundup.R, R/WL_roundup.R
\name{CT_roundup}
\alias{CT_roundup}
\title{Import, clean up, and export multiple raw conductivity-temperature, CT, data files from the HOBO CT Logger}
\usage{
CT_roundup(
  data.path,
  output.path,
  wl.serial = FALSE,
  tf_write = FALSE,
  tf_recursive = FALSE
)

CT_roundup(
  data.path,
  output.path,
  wl.serial = FALSE,
  tf_write = FALSE,
  tf_recursive = FALSE
)
}
\arguments{
\item{data.path}{Path to the input files}

\item{output.path}{Path for the output files}

\item{wl.serial}{Character string containing serial ID for a particular probe.  Used to filter from full dataset before returning final dataframe}

\item{tf_write}{Logical parameter indicating whether to save output files in an output folder. No default.}

\item{tf_recursive}{Logical parameter indicating whether to search within folders at the data.path. Default = FALSE}

\item{ct.serial}{Character string containing serial ID for a particular probe.  Used to filter from full dataset before returning final dataframe}

\item{ct.pattern}{Character string or object referencing a character string that identifies CT logger data within the specified file path}
}
\value{
For every imported CT data file, one tidied file with temperature-compensated conductance is exported and returned

For every imported CT data file, one tidied file with temperature-compensated conductance is exported and returned
}
\description{
This function loads all raw .csv files within the data.path folder.
It removes the first row of each file containing logger information, then retains column headers.
Unnecessary columns are removed, and the remaining columns are renamed as needed for simplicity and consistency.
Any rows containing NA's are dropped.

This function loads all raw .csv files within the data.path folder.
It removes the first row of each file containing logger information, then retains column headers.
Unnecessary columns are removed, and the remaining columns are renamed as needed for simplicity and consistency.
Any rows containing NA's are dropped.
}
