% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_cleanup.R
\name{CT_cleanup}
\alias{CT_cleanup}
\title{Clean up raw conductivity-temperature, CT, data file from the HOBO CT Logger}
\usage{
CT_cleanup(path, ct_serial, tf_write = FALSE, recursive_tf = FALSE)
}
\arguments{
\item{path}{Path to the input files}

\item{ct_serial}{Logger serial number used in naming the input file}

\item{tf_write}{Logical parameter indicating whether to save output files in an output folder. No default.}

\item{recursive_tf}{Logical parameter indicating whether to search within folders at the file path. Default = FALSE}

\item{output.path}{Path for the output files}
}
\value{
A cleaned dataframe of CT logger data
}
\description{
This function loads a raw CT file. It searches a directory for the correct logger's serial number,
plucks out the correct file or files and stacks all related files into one loaded csv.
It removes the first row of each relevant file containing logger information, then retains column headers.
Unnecessary columns are removed, and the remaining columns are renamed as needed for simplicity and consistency.
Any rows containing NA's are dropped.
}
