% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CT_two_cal.R
\name{CT_two_cal}
\alias{CT_two_cal}
\title{Calculate temperature-compensated Specific Conductance and two-point calibration for CT logger data}
\usage{
CT_two_cal(
  data,
  date,
  temp,
  Abs_pressure = 10,
  EC,
  high.Ref,
  low.Ref,
  startHigh,
  endHigh,
  startLow,
  endLow
)
}
\arguments{
\item{data}{The dataframe containing raw Electrical Conductivity values}

\item{date}{Date and time column used to filter out calibration times}

\item{temp}{Temperature at which electrical conductivity values were logged}

\item{Abs_pressure}{Absolute pressure, measured in dbar, used to calculate salinity using the PSS-78 equation.  Default = 10dbar.}

\item{EC}{Electrical conductivity logged at time of calibration}

\item{high.Ref}{The high range conductivity calibration solution specific conductance value, uS/cm at 25degC}

\item{low.Ref}{The low range conductivity calibration solution specific conductance value, uS/cm at 25degC}

\item{startHigh}{Date and time at the start of the high range calibration}

\item{endHigh}{Date and time at the end of the high range calibration}

\item{startLow}{Date and time at the start of the low range calibration}

\item{endLow}{Date and time at the end of the low range calibration}
}
\value{
The original dataframe with the calibrated Electrical Conductivity values of CT logger data
}
\description{
This function reads in a dataframe containing electrical conductivity logs from a conductivity-temperature logger
and calculates specific conductance from in situ electrical conductivity and temperature readings.
}
