% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pH_cleanup.R
\name{pH_cleanup}
\alias{pH_cleanup}
\title{Clean up raw pH data file from the HOBO pH Logger}
\usage{
pH_cleanup(
  data.path,
  output.path,
  path.pattern,
  tf.write = FALSE,
  tf.recursive = FALSE
)
}
\arguments{
\item{data.path}{Path to the input files}

\item{output.path}{Path for the output files}

\item{path.pattern}{Character string or object referencing a character string that identifies pH logger data within the specified file path}

\item{tf.write}{Logical parameter indicating whether to save output files in an output folder. Default = FALSE.}

\item{tf.recursive}{Logical parameter indicating whether to search within folders at the file path. Default = FALSE}

\item{pH.serial}{Logger serial number used in naming the input file}
}
\value{
For every imported pH data file, one tidied file is exported and returned
}
\description{
This function loads a raw pH file. It searches a directory for the correct logger's serial number,
plucks out the correct file or files and stacks all related files into one loaded csv.
It removes the first two rows of each relevant file containing logger information, then retains column headers.
Unnecessary columns are removed, and the remaining columns are renamed as needed for simplicity and consistency.
Any rows containing NA's are dropped.
}
